fix autonomy context
